// models/pokeAPI.ts
export interface PokeAPI {
  abilities: Ability[]
  base_experience: number
  cries: Cries
  forms: Species[]
  game_indices: GameIndex[]
  height: number
  id: number
  is_default: boolean
  location_area_encounters: string
  moves: NewMove[]
  name: string
  order: number
  species: Species
  sprites: Sprites
  stats: Stat[]
  types: Type[]
  weight: number
}

export interface Ability {
  ability: Species
  is_hidden: boolean
  slot: number
}

export interface AllAbilites {
  ability: Species
  is_hidden: boolean
  slot: number
}

export interface Species {
  name: string
  url: string
}

export interface Cries {
  latest: string
  legacy: string
}

export interface GameIndex {
  game_index: number
  version: Species
}

export interface NewMove {
  tag: string
  move: Species
  version_group_details: VersionGroupDetail[]
}

export interface VersionGroupDetail {
  level_learned_at: number
  move_learn_method: Species
  version_group: Species
}

export interface GenerationV {
  'black-white': Sprites
}

export interface GenerationIv {
  'diamond-pearl': Sprites
  'heartgold-soulsilver': Sprites
  platinum: Sprites
}

export interface Versions {
  'generation-i': GenerationI
  'generation-ii': GenerationIi
  'generation-iii': GenerationIii
  'generation-iv': GenerationIv
  'generation-v': GenerationV
  'generation-vi': { [key: string]: Home }
  'generation-vii': GenerationVii
  'generation-viii': GenerationViii
}

export interface Other {
  dream_world: DreamWorld
  home: Home
  'official-artwork': OfficialArtwork
  showdown: Sprites
}

export interface Sprites {
  back_default: string
  back_female: null
  back_shiny: string
  back_shiny_female: null
  front_default: string
  front_female: null
  front_shiny: string
  front_shiny_female: null
  other?: Other
  versions?: Versions
  animated?: Sprites
}
// ...
export interface Stat {
  base_stat: number
  effort: number
  stat: Species
}

export interface Type {
  slot: number
  type: Species
}

// Nature //

export interface NatureAPI {
  count: number
  next: string
  previous: null
  results: Result[]
}

export interface Result {
  name: string
  url: string
}

export interface NewNature {
  decreased_stat: DecreasedStat | null
  hates_flavor: DecreasedStat | null
  id: number
  increased_stat: DecreasedStat | null
  likes_flavor: DecreasedStat | null
  name: string
}

export interface DecreasedStat {
  name: string
  url: string
}

export interface Name {
  language: DecreasedStat
  name: string
}

// Ability //
// Define the types for PokeAPI response and transformed data
export interface NewAbility {
  flavor_text_entries: FlavorTextEntry[]
  name: string
}

// Move //
export interface NewCurrentMove {
  accuracy: number
  damage_class: ContestType
  id: number
  name: string
  power: number
  pp: number
  target: ContestType
  type: ContestType
}

export interface ContestType {
  name: string
  url: string
}

export interface FlavorTextEntry {
  flavor_text: string
  language: ContestType
  version_group: ContestType
}